remove Backtester object and implement backtesting in a funciton
