+++ examples/dmi_strategy.py
@@ -1,6 +1,6 @@
 import typing as t
 from datetime import datetime
-from backintime import TradingStrategy, Backtester
+from backintime import TradingStrategy, run_backtest
 from backintime.timeframes import Timeframes as tf
 from backintime.indicator_params import DMI
 from backintime.analyser.indicators.dmi import DMIResultSequence
@@ -19,12 +19,12 @@ def dmi_sell_signal(dmi: DMIResultSequence) -> bool:
 
 class DMIStrategy(TradingStrategy):
     """
-    Example DMI strategy.
+    Sample DMI strategy.
 
     Buy if +DI > -DI and ADX increases.
     Sell if +DI < -DI or ADX decreases.
     """
-    title = "Example DMI strategy"
+    title = "Sample DMI strategy"
     indicators = { DMI(tf.H4) }
 
     def tick(self):
@@ -40,9 +40,9 @@ feed = BinanceCandlesFactory('BTCUSDT', tf.M1)
 since = datetime.fromisoformat("2020-01-01 00:00+00:00")
 until = datetime.fromisoformat("2021-01-01 00:00+00:00")
 
-backtester = Backtester(DMIStrategy, feed)
-result = backtester.run(10_000, since, until, 
-                        maker_fee='0.005', taker_fee='0.005')
+result = run_backtest(DMIStrategy, feed,
+                      10_000, since, until, 
+                      maker_fee='0.005', taker_fee='0.005')
 print(result)
 print(result.get_stats('FIFO'))
 result.export()
